+++ src/controllers/categoryController.ts
@@ -1,7 +1,11 @@
 import { IReqUser, IPaginationQuery } from "../utils/interfaces";
 import { Response } from "express";
-import CategoryModel, { categoryDAO, TypeCategory } from "../models/categoryModel";
+import CategoryModel, {
+  categoryDAO,
+  TypeCategory,
+} from "../models/categoryModel";
 import * as response from "../utils/response";
+import { isValidObjectId } from "mongoose";
 
 export async function create(req: IReqUser, res: Response) {
   try {
@@ -18,13 +22,13 @@ export async function create(req: IReqUser, res: Response) {
 }
 
 export async function findAll(req: IReqUser, res: Response) {
-   const {
+  const {
     page = 1,
     limit = 10,
     search,
   } = req.query as unknown as IPaginationQuery;
   try {
-     // handling search
+    // handling search
     const query = {};
 
     if (search) {
@@ -37,19 +41,24 @@ export async function findAll(req: IReqUser, res: Response) {
     }
 
     const result = await CategoryModel.find(query)
-        .limit(limit)
-        .skip((page - 1) * limit) // for skip data example page 2 limit 10, so skip (2-1)*10 = 10 so data is start from 11
-        .sort({ createdAt: -1}) // sort by createdAt descending
-        .exec();
+      .limit(limit)
+      .skip((page - 1) * limit) // for skip data example page 2 limit 10, so skip (2-1)*10 = 10 so data is start from 11
+      .sort({ createdAt: -1 }) // sort by createdAt descending
+      .exec();
 
-    // for count all data which match with query example with search  
+    // for count all data which match with query example with search
     const count = await CategoryModel.countDocuments(query);
 
-    response.pagination(res, result, {
+    response.pagination(
+      res,
+      result,
+      {
         totalPages: Math.ceil(count / limit), // total all data / limit for get total pages
         currentPage: page,
         total: count, // total all data which match with query
-    }, 'Success get all categories');
+      },
+      "Success get all categories",
+    );
   } catch (error) {
     response.error(res, error, "failed find all category");
   }
@@ -59,8 +68,17 @@ export async function findOne(req: IReqUser, res: Response) {
   try {
     const { id } = req.params;
 
+    if (!isValidObjectId(id)) {
+      return response.badRequest(res, "id is not valid, please check your id!");
+    }
+
     const result = await CategoryModel.findById(id);
-    response.success(res, result, 'Success get one category');
+
+    if (!result) {
+      response.notFound(res, "Category not found");
+    }
+
+    response.success(res, result, "Success get one category");
   } catch (error) {
     response.error(res, error, "failed find one category");
   }
@@ -69,11 +87,21 @@ export async function findOne(req: IReqUser, res: Response) {
 export async function update(req: IReqUser, res: Response) {
   try {
     const { id } = req.params;
+
+    if (!isValidObjectId(id)) {
+      return response.badRequest(res, "id is not valid, please check your id!");
+    }
+
     const payload = req.body as Partial<TypeCategory>;
 
     const result = await CategoryModel.findByIdAndUpdate(id, payload, {
       new: true,
     });
+
+    if (!result) {
+      response.notFound(res, "Category not found");
+    }
+
     response.success(res, result, "Success update a category");
   } catch (error) {
     response.error(res, error, "failed update category");
@@ -84,9 +112,18 @@ export async function remove(req: IReqUser, res: Response) {
   try {
     const { id } = req.params;
 
+    if (!isValidObjectId(id)) {
+      return response.badRequest(res, "id is not valid, please check your id!");
+    }
+
     const result = await CategoryModel.findByIdAndDelete(id, {
       new: true,
     });
+
+    if (!result) {
+      response.notFound(res, "Category not found");
+    }
+    
     response.success(res, result, "Success remove a category");
   } catch (error) {
     response.error(res, error, "failed remove category");
